+++ tests/testthat/test-breaks.R
@@ -125,6 +125,18 @@ test_that("brk_equally", {
 })
 
 
+test_that("brk_pretty", {
+  expect_silent(brks <- brk_res(brk_pretty(5), x = 1:10))
+  expect_equivalent(brks, brk_res(brk_default(pretty(1:10)), x = 1:10))
+
+  expect_silent(brks2 <- brk_res(brk_pretty(5, high.u.bias = 0), x = 1:10))
+  expect_equivalent(
+    brks2,
+    brk_res(brk_default(pretty(1:10, high.u.bias = 0)), x = 1:10)
+  )
+})
+
+
 test_that("printing", {
   b <- brk_default(1:3)
   expect_output(print(b))